+++ doc/conf.py
@@ -43,7 +43,7 @@ if on_rtd:
 # -- General configuration ------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
-needs_sphinx = '1.6.1'
+needs_sphinx = '3.0.0'
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
@@ -59,9 +59,9 @@ extensions = [
     'sphinx.ext.ifconfig',
     'sphinx.ext.viewcode',
     'sphinx.ext.autosummary',
-    'sphinx.ext.napoleon',
 
     'sphinxext.inline_code_highlight',
+    'numpydoc',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -78,7 +78,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'Mizani'
-copyright = '2016, Hassan Kibirige'
+copyright = '2020, Hassan Kibirige'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -114,7 +114,9 @@ release = version
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
-exclude_patterns = []
+exclude_patterns = [
+    '_build'
+]
 
 # The reST default role (used for this markup: `text`) to use for all
 # documents.
@@ -328,8 +330,8 @@ texinfo_documents = [
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
     'python': ('https://docs.python.org/3/', None),
-    'matplotlib': ('http://matplotlib.org/', None),
-    'numpy': ('https://docs.scipy.org/doc/numpy', None),
+    'matplotlib': ('https://matplotlib.org/', None),
+    'numpy': ('https://numpy.org/doc/stable', None),
     'scipy': ('https://docs.scipy.org/doc/scipy/reference', None),
 }
 
@@ -342,6 +344,32 @@ extlinks = {
     'issue': ('https://github.com/has2k1/mizani/issues/%s', 'GH')
 }
 
+# numpydoc
+numpydoc_show_class_members = False
+numpydoc_class_members_toctree = False
+numpydoc_xref_param_type = True
+numpydoc_xref_aliases = {
+    # python
+    'sequence': ':term:`python:sequence`',
+    'iterable': ':term:`python:iterable`',
+    'string': 'str',
+    'tuples': 'tuple',
+    'boolean': 'bool',
+    # numpy
+    'array': 'numpy.ndarray',
+    'np.array': 'numpy.ndarray',
+    'ndarray': 'numpy.ndarray',
+    'array-like': ':term:`array-like<numpy:array_like>`',
+    'array_like': ':term:`numpy:array_like`',
+    # pandas
+    'dataframe': 'pandas.DataFrame',
+    'DataFrame': 'pandas.DataFrame',
+    'Series': 'pandas.Series',
+    'series': 'pandas.Series',
+}
+numpydoc_xref_ignore = {'type', 'optional', 'default'}
+
+
 def setup(app):
-    app.add_javascript('copybutton.js')
-    app.add_stylesheet('custom.css')
+    app.add_js_file('copybutton.js')
+    app.add_css_file('custom.css')