+++ mizani/tests/test_transforms.py
@@ -9,11 +9,13 @@ import pytest
 from mizani.breaks import mpl_breaks, minor_breaks
 from mizani.transforms import (
     trans,
-    asn_trans, atanh_trans, boxcox_trans, datetime_trans,
+    asn_trans, atanh_trans, boxcox_trans, modulus_trans,
+    datetime_trans,
     exp_trans, identity_trans, log10_trans, log1p_trans,
     log2_trans, log_trans, probability_trans, reverse_trans,
-    sqrt_trans, timedelta_trans, pd_timedelta_trans, trans_new,
-    gettrans)
+    sqrt_trans, timedelta_trans, pd_timedelta_trans,
+    pseudo_log_trans, reciprocal_trans,
+    trans_new, gettrans)
 
 arr = np.arange(1, 100)
 
@@ -63,7 +65,7 @@ def _test_trans(trans, x):
     t = gettrans(trans())
     xt = t.transform(x)
     x2 = t.inverse(xt)
-    is_log_trans = (t.__class__.__name__.startswith('log') and
+    is_log_trans = ('log' in t.__class__.__name__ and
                     hasattr(t, 'base'))
     # round trip
     npt.assert_allclose(x, x2)
@@ -94,8 +96,29 @@ def test_atanh_trans():
 
 
 def test_boxcox_trans():
-    _test_trans(boxcox_trans(0), arr)
     _test_trans(boxcox_trans(0.5), arr*10)
+    _test_trans(boxcox_trans(1), arr)
+    with pytest.raises(ValueError):
+        x = np.arange(-4, 4)
+        _test_trans(boxcox_trans(0.5), x)
+
+    # Special case, small p and p = 0
+    with pytest.warns(RuntimeWarning):
+        _test_trans(boxcox_trans(1e-8), arr)
+        _test_trans(boxcox_trans(0), arr)
+
+    x = [0, 1, 2, 3]
+    t = boxcox_trans(0)
+    with pytest.warns(RuntimeWarning):
+        xt = t.transform(x)
+    xti = t.inverse(xt)
+    assert np.isneginf(xt[0])
+    npt.assert_array_almost_equal(x, xti)
+
+
+def test_modulus_trans():
+    _test_trans(modulus_trans(0), arr)
+    _test_trans(modulus_trans(0.5), arr*10)
 
 
 def test_exp_trans():
@@ -141,6 +164,23 @@ def test_logn_trans():
                            breaks=mpl_breaks(),
                            minor_breaks=minor_breaks())
     _test_trans(log4_trans, arr)
+
+
+def test_reciprocal_trans():
+    x = np.arange(10, 21)
+    _test_trans(reciprocal_trans, x)
+
+
+def test_pseudo_log_trans():
+    p = np.arange(-4, 4)
+    pos = [10 ** int(x) for x in p]
+    arr = np.hstack([-np.array(pos[::-1]), pos])
+    _test_trans(pseudo_log_trans, arr)
+    _test_trans(pseudo_log_trans(base=16), arr)
+    _test_trans(
+        pseudo_log_trans(base=10, minor_breaks=minor_breaks(n=5)),
+        arr
+    )
 
 
 def test_probability_trans():